requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
fairscale==0.4.13
filelock==3.14.0
fonttools==4.57.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jmespath==0.10.0
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
matplotlib==3.7.5
mdurl==0.1.2
mmcv==2.1.0
mmengine==0.10.7
mmdet==3.3.0
model-index==0.1.11
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.22.0
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
requests==2.28.2
rich==13.4.2
scipy==1.10.1
shapely==2.0.7
six==1.17.0
tabulate==0.9.0
termcolor==2.4.0
terminaltables==3.1.10
tomli==2.2.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==1.26.20
yapf==0.43.0
zipp==3.20.2